==> Red_Alerts_Analyze/3. Visualization.ipynb <==
<h1><center><font color='Black'>Visualization</font></center></h1>

#-------------Imports-----------#

import pandas as pd
import json
from datetime import date
import datetime as dt
import numpy as np
import matplotlib.pyplot as plt
from matplotlib import pyplot as plt
from wordcloud import WordCloud
from bidi.algorithm import get_display
from PIL import Image
import seaborn as sns

todays_date = date.today() #Get Today'S Date
curryear = todays_date.year
swipe = curryear - 2014 #Start From Curr Year To 2014

colors = ['#ff9999','#66b3ff','#99ff99','#ffcc99' ,'#abac9a' ]

df = pd.read_csv('Data/df.csv', encoding='utf-8-sig' )
df2 = pd.read_csv('Data/df2.csv', encoding='utf-8-sig')
df

Change the format of the DateTime into "YY-MM-DD"


df["Date"] = pd.to_datetime(df['Date'],format= '%Y-%m-%d')



df['Datetime'] = pd.to_datetime(df['Date'].apply(str)+' '+df['Time'])


Sum all the alerts per year

Years =[year for year in range(2014,curryear+1)]
values =[df['Date'][df.Date.dt.year == year ].count() for year in Years ]

for year in range(2014,curryear+1):
    num = df['Date'][df.Date.dt.year== year ].count()
    print("Number Of Alerts At",year,"------>", num)

df_alerst_per_year = pd.DataFrame(data = { "Year" : Years, "Count" : values } )

We will now present in bar plot the amount of alerts for each year

# Double click to see the code

plt.figure(figsize=[15, 7])

# Creating a bar chart with bars of different color
plt.bar(df_alerst_per_year.Year, df_alerst_per_year.Count, width=0.7, edgecolor='blueviolet',
        color=['r', 'y', 'g', 'b', 'c', 'k', 
               'olive', 'gray', 'pink', 'maroon'], linewidth=2)

plt.title('The number of alarms each year', fontsize=17)
plt.xlabel('Years', fontsize=15)
plt.ylabel('Number of alerts', fontsize=15)
plt.show()




Now let's see what the hour with the most alerts from 2014 to present

Hours =[hour for hour in range(0,24)]
values =[df['Datetime'][df.Datetime.dt.hour== hour ].count() for hour in Hours]

df_common_hour = pd.DataFrame(data = { "Hour" : Hours, "Count" : values } )

# Double click to see the code
for hour in range(0,24):
    num = df['Datetime'][df.Datetime.dt.hour== hour ].count()
    #print("Number Of Alerts At",hour,"------>", num)

plt.figure(figsize=[15, 7])

# Creating a bar chart with bars of different color
plt.bar(df_common_hour['Hour'],df_common_hour['Count'] ,width=0.7, edgecolor='blueviolet', color=['r', 'y', 'g', 'b', 'c', 'k', 'olive', 'gray', 'pink', 'maroon'], linewidth=2)

plt.title('The hour with the most alerts from 2014 to present', fontsize=15)
plt.xlabel('Hour', fontsize=15)
plt.ylabel('Number of alerts', fontsize=15)
plt.show()


We see that around 10 AM is the result, now let's see it more clearly in the form of a clock

# Double click to see the code
colors2 = []
for value in values:
    if value > 2000:
        colors2.append("#ff9999") # Red
    elif value < 2000 and value > 1000:
        colors2.append("#66b3ff") #Blue
    else:
        colors2.append("#99ff99") #Green

        

fig = plt.figure(figsize=(10,10))

ax = plt.subplot(111, projection = 'polar')

ax.bar(df_common_hour['Hour']-1, df_common_hour['Count'],  width = 0.29, alpha=0.8, color=colors2, edgecolor='blueviolet', linewidth=1)
ax.bar(0, 0 , color='#ff9999',label = "More Then 2000 Rockets")
ax.bar(0, 0 , color='#66b3ff',label = "Between 1000 And 2000 Rockets")
ax.bar(0, 0 , color='#99ff99',label = "Less Then  1000 Rockets")
# Make the labels go clockwise
ax.set_theta_direction(-1)

#Place Zero at Top
ax.set_theta_offset(np.pi/2)

#Set the circumference ticks
ax.set_xticks(np.linspace(0,2*np.pi, 24, endpoint=False))

# set the label names
#ticks = ['12 AM', '1 AM', '2 AM', '3 AM', '4 AM', '5 AM', '6 AM', '7 AM','8 AM','9 AM','10 AM','11 AM','12 PM', '1 PM', '2 PM', '3 PM', '4 PM',  '5 PM', '6 PM', '7 PM', '8 PM', '9 PM', '10 PM', '11 PM' ]
ticks = ['00:00', '01:00', '02:00', '03:00', '04:00', '05:00', '06:00', '07:00','08:00','09:00','10:00','11:00','12:00', '13:00', '14:00', '15:00', '16:00',  '17:00', '18:00', '19:00', '20:00', '21:00', '22:00', '23:00' ]
ax.set_xticklabels(ticks)

# suppress the radial labels
plt.setp(ax.get_yticklabels(), visible=False)

#Bars to the wall
plt.ylim(0,max(values))

plt.legend(bbox_to_anchor=(1,0), fancybox=True, shadow=True)
plt.show()

Now let's see the cities with to most alerts from 2014 to present, in the first plot we see the top 10, the second one we used *WordCloud* to see the cities with the most alerts in form of text

# Double click to see the code

col_map = plt.get_cmap('Paired')
text = df.loc[:,'Zone'].apply(lambda x: x[::-1]).value_counts()[:30]
plt.title('The cities with the most alerts from 2014 to present (Top 10)', fontsize=15)
df.loc[:,'Zone'].apply(lambda x: x[::-1]).value_counts()[:10].plot(kind='bar', width=0.7,color=col_map.colors,edgecolor='k',linewidth=2,figsize=[15, 7])
plt.xlabel('Cities', fontsize=15)
plt.ylabel('Number of alerts', fontsize=15)


def transform_format(val):
    if val != 0:
        return 255
    else:
        return val

    
    
image_file=Image.open("Photos/rocket.png")
image_file=image_file.convert('1')
rocket_mask = np.array(image_file)

transformed_rocket_mask = np.ndarray((rocket_mask.shape[0],rocket_mask.shape[1]), np.int32)

for i in range(len(rocket_mask)):
    transformed_rocket_mask[i] = list(map(transform_format, rocket_mask[i]))

wc = WordCloud(background_color="White", max_words=100, mask=transformed_rocket_mask,contour_width=3, contour_color='gray', font_path= 'Data/FreeSansBold.ttf')

wc=wc.generate_from_frequencies(text)
plt.figure(figsize=[20,10])
plt.imshow(wc, interpolation='bilinear')
plt.axis("off")
plt.show()



Let's compare the big cities

# Double click to see the code

cities = ['באר שבע' , 'אשדוד'  , 'תל אביב'  , 'אשקלון'  , 'ראשון לציון'  ]

values = []
for city in cities:
    num = df["Zone"][df.Zone.str.contains(city)].count()
    values.append(num)

cities = [x[::-1] for x in cities]   

fig, ax = plt.subplots(figsize=(7, 6))
ax.pie(values, labels=cities, colors=colors, autopct='%1.1f%%',explode=(0,0,0, 0.1, 0) , shadow=True , startangle=270 , textprops={'fontsize': 14})

ax.set_title("The big cities with the most alerts from 2014-present",fontsize=18)
plt.tight_layout()
